fix paste phone number issue autoFill is show when focus on textField
